+++ api/src/main/java/org/pmiops/workbench/access/AccessTierSyncService.java
@@ -1,0 +1,12 @@
+package org.pmiops.workbench.access;
+
+import org.pmiops.workbench.actionaudit.Agent;
+import org.pmiops.workbench.db.model.DbUser;
+
+public interface AccessTierSyncService {
+
+  /**
+   * Ensures that the data access tiers for the user reflect the state of other fields on the user
+   */
+  DbUser updateUserAccessTiers(DbUser dbUser, Agent agent);
+}
