+++ api/src/main/java/org/pmiops/workbench/api/CloudTaskUserController.java
@@ -120,7 +120,6 @@ public class CloudTaskUserController implements CloudTaskUserApiDelegate {
       DbUser user = userDao.findUserByUserId(userId);
 
       try {
-        user = userService.syncDuccVersionStatus(user, Agent.asSystem());
 
         // 2FA synchronization requires an outgoing call to gsuite. For this reason, we
         // optimize to only verify that users who have 2FA enabled, still have it enabled.
@@ -135,11 +134,11 @@ public class CloudTaskUserController implements CloudTaskUserApiDelegate {
           }
         }
 
-        // Always synchronize for consistency. Under normal operation only the passage of time
-        // should result in access expiration changes here, but this is also a backstop for ensuring
-        // the database is consistent with our access tier groups, e.g. due to partial system
-        // failures or bugs.
-        userService.updateUserAccessTiers(user, Agent.asSystem());
+        // Note: each module synchronization calls updateUserAccessTiers() which checks the status
+        // of *all* modules, so this serves as a general fallback as well (e.g. due to partial
+        // system failures or bugs), ensuring that the database and access tier groups are
+        // consistent with access module statuses.
+        user = userService.syncDuccVersionStatus(user, Agent.asSystem());
       } catch (WorkbenchException e) {
         log.log(Level.SEVERE, "failed to synchronize access for user " + user.getUsername(), e);
         errorCount++;
