+++ api/src/test/java/org/pmiops/workbench/testconfig/UserServiceTestConfiguration.java
@@ -1,6 +1,7 @@
 package org.pmiops.workbench.testconfig;
 
 import org.pmiops.workbench.access.AccessModuleNameMapperImpl;
+import org.pmiops.workbench.access.AccessTierSyncServiceImpl;
 import org.pmiops.workbench.db.dao.UserServiceImpl;
 import org.pmiops.workbench.institution.InstitutionMapperImpl;
 import org.pmiops.workbench.institution.InstitutionServiceImpl;
@@ -23,5 +24,6 @@ import org.springframework.context.annotation.Import;
   PublicInstitutionDetailsMapperImpl.class,
   InstitutionTierConfigMapperImpl.class,
   AccessModuleNameMapperImpl.class,
+  AccessTierSyncServiceImpl.class,
 })
 public class UserServiceTestConfiguration {}
